1.change loadLibrary;2.range change;3.fix maxlines bug;4.registe AniTextGlobal;

+++ frameworks/text/interface/mlb/ani/include/ani_text_rect_converter.h
@@ -24,8 +24,8 @@
 
 namespace OHOS::Text::ANI {
 struct RectRange {
-    size_t start;
-    size_t end;
+    int64_t start;
+    int64_t end;
 };
 
 class AniTextRectConverter {
